# openapi.yml
openapi: '3.0.2'
info:
  title: Node Starter
  version: '1.0'

servers:
  - url: http://localhost:3000
tags:
  - name: main
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /home:
    get:
      summary: "Hello World text example"
      tags: [main]
      responses:
        '200':
          description: "Just a Hello World"
          content:
            text/plain:
              schema:
                type: string
                example: "Hello World"

  /json:
    get:
      summary: "JSON Example"
      tags: [main]
      responses:
        '200':
          description: "A JSON response"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OK"
  /get-request-body:
    post:
      summary: "It will show all request body properties"
      tags: [main]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
            example: {"name": "Carlos"}
          
      responses:
        '200':
          description: "Show response body"
          content:
            application/json:
              schema:
                type: object
                properties: 
                  body:
                    type: object
                    example: {"name": "Carlos"}
  /get-auth:
    post:
      summary: "It will show token"
      tags: [main]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Show token"
          content:
            application/json:
              schema:
                type: object
                properties: 
                  token:
                    type: object
                    example: 00x1
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: 
                    type: object
                    example: "You are not authenticated"
                    